requirements.txt
# Core Dependencies
psycopg2-binary==2.9.9
pgvector==0.2.4
numpy==1.26.4
sentence-transformers==2.5.1

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.11.0

# Ollama Integration
ollama==0.1.6
requests==2.31.0

# CLI
typer==0.9.0
rich==13.7.0
prompt-toolkit==3.0.43

# GUI
customtkinter==5.2.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1

# Vector Operations
faiss-cpu==1.7.4
